+++ ingenialink/bitfield.py
@@ -1,4 +1,4 @@
-from typing import Dict, Iterable
+from collections.abc import Iterable
 
 
 def _bit_mask(selected_bits: Iterable[int]) -> int:
@@ -31,6 +31,7 @@ class BitField:
         self._mask = _bit_mask(range(start, end + 1))
 
     def __eq__(self, other: object) -> bool:
+        """Compare bitfields."""
         if not isinstance(other, BitField):
             return False
         return (self._start == other._start) and (self._end == other._end)
@@ -41,7 +42,7 @@ class BitField:
         return cls(start=bit, end=bit)
 
     @staticmethod
-    def parse_bitfields(bitfields: Dict[str, "BitField"], value: int) -> Dict[str, int]:
+    def parse_bitfields(bitfields: dict[str, "BitField"], value: int) -> dict[str, int]:
         """Parse value into bitfields.
 
         Separates a integer value into a dictionary of values, where the key is the bitfield name
@@ -63,7 +64,7 @@ class BitField:
         }
 
     @staticmethod
-    def set_bitfields(bitfields: Dict[str, "BitField"], values: Dict[str, int], value: int) -> int:
+    def set_bitfields(bitfields: dict[str, "BitField"], values: dict[str, int], value: int) -> int:
         """Set bitfields to a value.
 
         Args: