+++ ingenialink/canopen/servo.py
@@ -1,8 +1,9 @@
-from typing import Any, Callable, List, Optional, Union
+from typing import Any, Callable, Optional, Union
 
 import canopen
 import ingenialogger
 from canopen.emcy import EmcyError
+from typing_extensions import override
 
 from ingenialink.canopen.register import CanopenRegister
 from ingenialink.constants import CAN_MAX_WRITE_SIZE
@@ -45,10 +46,11 @@ class CanopenServo(Servo):
         servo_status_listener: bool = False,
     ) -> None:
         self.__node = node
-        self.__emcy_observers: List[Callable[[EmergencyMessage], None]] = []
+        self.__emcy_observers: list[Callable[[EmergencyMessage], None]] = []
         self.__node.emcy.add_callback(self._on_emcy)
-        super(CanopenServo, self).__init__(target, dictionary_path, servo_status_listener)
+        super().__init__(target, dictionary_path, servo_status_listener)
 
+    @override
     def read(
         self, reg: Union[str, Register], subnode: int = 1, **kwargs: Any
     ) -> Union[int, float, str, bytes]:
@@ -94,7 +96,7 @@ class CanopenServo(Servo):
         finally:
             self._lock.release()
         if not isinstance(value, bytes):
-            return bytes()
+            return b""
         return value
 
     def emcy_subscribe(self, callback: Callable[[EmergencyMessage], None]) -> None:
@@ -117,6 +119,7 @@ class CanopenServo(Servo):
 
     def _on_emcy(self, emergency_msg: EmcyError) -> None:
         """Receive an emergency message from canopen and transform it to a EmergencyMessage.
+
         Afterward, send the EmergencyMessage to all the subscribed callbacks.
 
         Args:
@@ -138,11 +141,10 @@ class CanopenServo(Servo):
             return is_register_valid
         # Exclude the RxPDO and TxPDO related registers
         # Check INGK-980
-        if register.identifier is not None and register.identifier.startswith(
-            ("CIA301_COMMS_TPDO", "CIA301_COMMS_RPDO")
-        ):
-            return False
-        return True
+        return not (
+            register.identifier is not None
+            and register.identifier.startswith(("CIA301_COMMS_TPDO", "CIA301_COMMS_RPDO"))
+        )
 
     @staticmethod
     def _monitoring_disturbance_map_can_address(address: int, subnode: int) -> int: