+++ tests/eoe/test_eoe_network.py
@@ -20,7 +20,7 @@ def connect(read_config):
 
 
 @pytest.mark.eoe
-def test_eoe_connection(connect_to_slave, read_config):
+def test_eoe_connection(connect_to_slave):
     eoe_service_ip = "127.0.0.1"
     eoe_service_port = 8888
     servo, net = connect_to_slave
@@ -70,7 +70,7 @@ def test_eoe_disconnection(connect):
 )
 @pytest.mark.eoe
 def test_eoe_command_msg(cmd, data):
-    data_bytes = bytes() if data is None else data
+    data_bytes = b"" if data is None else data
     data_filling = EoENetwork.NULL_TERMINATOR * (EoENetwork.EOE_MSG_DATA_SIZE - len(data_bytes))
     msg = EoENetwork._build_eoe_command_msg(cmd.value, data_bytes)
     cmd_field = msg[: EoENetwork.EOE_MSG_CMD_SIZE]