+++ virtual_drive/core.py
@@ -5,11 +5,12 @@ import socket
 import time
 from enum import Enum, IntEnum
 from threading import Thread, Timer
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any, Optional, Union
 
 import numpy as np
 from numpy.typing import NDArray
 from scipy import signal
+from typing_extensions import override
 
 from ingenialink.constants import ETH_BUF_SIZE, MONITORING_BUFFER_SIZE
 from ingenialink.dictionary import Interface
@@ -24,7 +25,9 @@ from virtual_drive import constants
 
 from .environment import Environment
 from .gpios import Gpios
-from .signals import Signal
+
+if TYPE_CHECKING:
+    from .signals import Signal
 
 R_VALUE = 1.1
 L_VALUE = 3.9e-4
@@ -38,13 +41,15 @@ INC_ENC1_RESOLUTION = 4096
 INC_ENC2_RESOLUTION = 2000
 
 
-class MSG_TYPE(Enum):
+class MsgType(Enum):
+    """Message type enum."""
+
     RECEIVED = "RECEIVED"
     SENT = "SENT"
 
 
 class OperationMode(IntEnum):
-    """Operation Mode Enum"""
+    """Operation Mode Enum."""
 
     VOLTAGE = 0x00
     CURRENT_AMPLIFIER = 0x01
@@ -65,7 +70,7 @@ class OperationMode(IntEnum):
 
 
 class PhasingMode(IntEnum):
-    """Phasing modes"""
+    """Phasing modes."""
 
     NON_FORCED = 0
     """Non forced"""
@@ -76,7 +81,7 @@ class PhasingMode(IntEnum):
 
 
 class SensorType(IntEnum):
-    """Summit series feedback type enum"""
+    """Summit series feedback type enum."""
 
     ABS1 = 1
     """Absolute encoder 1"""
@@ -95,7 +100,7 @@ class SensorType(IntEnum):
 
 
 class GeneratorMode(IntEnum):
-    """Generator modes"""
+    """Generator modes."""
 
     CONSTANT = 0
     """Constant"""
@@ -131,9 +136,9 @@ class BasePlant:
         self.drive = drive
         self.monitoring_frequency = VirtualMonitoring.FREQUENCY
         self.plant: signal.TransferFunction
-        self.set_point_register = self.drive.get_register(1, id=self.REGISTER_SET_POINT)
-        self.command_register = self.drive.get_register(1, id=self.REGISTER_COMMAND)
-        self.value_register = self.drive.get_register(1, id=self.REGISTER_VALUE)
+        self.set_point_register = self.drive.get_register(1, register_id=self.REGISTER_SET_POINT)
+        self.command_register = self.drive.get_register(1, register_id=self.REGISTER_COMMAND)
+        self.value_register = self.drive.get_register(1, register_id=self.REGISTER_VALUE)
 
     def emulate_plant(self, from_disturbance: bool = True) -> None:
         """Emulate the plant by filtering the excitation signal with the plant's frequency response.
@@ -166,17 +171,18 @@ class BasePlant:
 
     def _filter_signal(
         self,
-        input_signal: NDArray[np.float_],
+        input_signal: NDArray[np.float64],
         use_fft_method: bool = True,
         initial_value: float = 0.0,
         plant: Optional[signal.TransferFunction] = None,
-    ) -> NDArray[np.float_]:
+    ) -> NDArray[np.float64]:
         """Filter signal with the plant's frequency response.
 
         Args:
             input_signal: Signal to be filtered.
             use_fft_method: If True the FFT method is used.
             initial_value: Initial value.
+            plant: Plant's transfer function.
 
         Returns:
             Filtered signal.
@@ -245,7 +251,7 @@ class BaseOpenLoopPlant(BasePlant):
         plant: Plant's discrete transfer function.
     """
 
-    def __init__(self, drive: "VirtualDrive", num: List[float], den: List[float]) -> None:
+    def __init__(self, drive: "VirtualDrive", num: list[float], den: list[float]) -> None:
         super().__init__(drive)
         self.plant = signal.TransferFunction(num, den).to_discrete(dt=1 / self.monitoring_frequency)
         self.drive.reg_noise_amplitude[self.REGISTER_VALUE] = 0.01
@@ -301,6 +307,7 @@ class BaseClosedLoopPlant(BasePlant):
 
         self.plant = signal.TransferFunction(num, den, dt=1 / self.monitoring_frequency)
 
+    @override
     def emulate_plant(self, from_disturbance: bool = True) -> None:
         self.create_closed_loop_plant()
         super().emulate_plant(from_disturbance=from_disturbance)
@@ -375,6 +382,7 @@ class PlantClosedLoopRL(BaseClosedLoopPlant):
     TUNING_OPERATION_MODE = OperationMode.CURRENT
     OL_PLANT_INPUT = PlantOpenLoopRL.REGISTER_COMMAND
 
+    @override
     def emulate_plant(self, from_disturbance: bool = True) -> None:
         super().emulate_plant(from_disturbance)
 
@@ -728,8 +736,8 @@ class VirtualInternalGenerator:
         n_samples = int(VirtualMonitoring.FREQUENCY * period * self.cycles)
         time_vector = self.start_time + np.linspace(0, period * self.cycles, n_samples)
 
-        signal_period: NDArray[np.float_] = self.offset + np.linspace(
-            0, self.gain, int(n_samples / self.cycles), dtype=np.float_
+        signal_period: NDArray[np.float64] = self.offset + np.linspace(
+            0, self.gain, int(n_samples / self.cycles), dtype=np.float64
         )
         pos_signal = np.tile(signal_period, self.cycles)
         initial_value = self.drive.get_value_by_id(1, self.ACTUAL_POSITION_REGISTER)
@@ -748,7 +756,7 @@ class VirtualInternalGenerator:
         self.drive.reg_signals[self.ACTUAL_POSITION_REGISTER] = pos_signal
         self.drive.reg_time[self.ACTUAL_POSITION_REGISTER] = time_vector
 
-    def __create_halls_encoder_signal(self, pos_signal: NDArray[np.float_]) -> NDArray[np.float_]:
+    def __create_halls_encoder_signal(self, pos_signal: NDArray[np.float64]) -> NDArray[np.float64]:
         """Create the halls encoder signal by discretizing the pos_signal using the hall values.
 
         Args:
@@ -773,7 +781,7 @@ class VirtualInternalGenerator:
             encoder_signal[sample_ix] = self.HALL_VALUES[hall_value_ix]
         return encoder_signal
 
-    def __create_abs_encoder_signal(self, pos_signal: NDArray[np.float_]) -> NDArray[np.float_]:
+    def __create_abs_encoder_signal(self, pos_signal: NDArray[np.float64]) -> NDArray[np.float64]:
         """Crete the absolute encoder signal by saturating the position signal.
 
         Args:
@@ -873,12 +881,12 @@ class VirtualMonDistBase:
         self.enabled = False
         self.disable()
         self.number_mapped_registers = 0
-        self.channels_data: Dict[int, NDArray[np.float_]] = {}
-        self.channels_dtype: Dict[int, REG_DTYPE] = {}
-        self.channels_address: Dict[int, int] = {}
-        self.channels_subnode: Dict[int, int] = {}
-        self.channels_size: Dict[int, int] = {}
-        self.channels_signal: Dict[int, NDArray[np.float_]] = {}
+        self.channels_data: dict[int, NDArray[np.float_]] = {}
+        self.channels_dtype: dict[int, REG_DTYPE] = {}
+        self.channels_address: dict[int, int] = {}
+        self.channels_subnode: dict[int, int] = {}
+        self.channels_size: dict[int, int] = {}
+        self.channels_signal: dict[int, NDArray[np.float_]] = {}
 
     def enable(self) -> None:
         """Enable Monitoring/Disturbance."""
@@ -958,7 +966,7 @@ class VirtualMonDistBase:
     def available_bytes(self, n_bytes: int) -> None:
         self.drive.set_value_by_id(0, self.AVAILABLE_BYTES_REG, n_bytes)
 
-    def get_mapped_register(self, channel: int) -> Tuple[int, int, int, int]:
+    def get_mapped_register(self, channel: int) -> tuple[int, int, int, int]:
         """Decodes the register with the information of a mapped register.
 
         Args:
@@ -1040,6 +1048,7 @@ class VirtualMonitoring(VirtualMonDistBase):
         self.start_time = 0.0
         super().__init__(drive)
 
+    @override
     def enable(self) -> None:
         super().map_registers()
         super().enable()
@@ -1049,6 +1058,7 @@ class VirtualMonitoring(VirtualMonDistBase):
         self.update_data()
 
     def update_data(self) -> None:
+        """Update the monitoring data."""
         self.__create_signals()
 
         # Store data
@@ -1087,7 +1097,8 @@ class VirtualMonitoring(VirtualMonDistBase):
                     )
                     self.channels_signal[channel] = self.channels_signal[channel][indexes]
                 if self.drive.reg_noise_amplitude[reg_id] > 0:
-                    noise = self.drive.reg_noise_amplitude[reg_id] * np.random.normal(
+                    rng = np.random.default_rng()
+                    noise = self.drive.reg_noise_amplitude[reg_id] * rng.normal(
                         size=len(self.channels_signal[channel])
                     )
                     self.channels_signal[channel] = self.channels_signal[channel] + noise
@@ -1107,7 +1118,7 @@ class VirtualMonitoring(VirtualMonDistBase):
 
     def _store_data_bytes(self) -> None:
         """Convert signals into a bytes and store it at MON_DATA register."""
-        byte_array = bytes()
+        byte_array = b""
         n_samples = len(self.channels_data[0])
         for sample in range(n_samples):
             for channel in range(self.number_mapped_registers):
@@ -1150,9 +1161,10 @@ class VirtualDisturbance(VirtualMonDistBase):
 
     def __init__(self, drive: "VirtualDrive") -> None:
         self.start_time = 0.0
-        self.received_bytes = bytes()
+        self.received_bytes = b""
         super().__init__(drive)
 
+    @override
     def enable(self) -> None:
         self.start_time = time.time()
         for channel in range(self.number_mapped_registers):
@@ -1162,8 +1174,9 @@ class VirtualDisturbance(VirtualMonDistBase):
             self.drive.reg_time[str(reg.identifier)] += self.start_time
         super().enable()
 
+    @override
     def disable(self) -> None:
-        self.received_bytes = bytes()
+        self.received_bytes = b""
         return super().disable()
 
     def append_data(self, data: bytes) -> None:
@@ -1187,8 +1200,8 @@ class VirtualDisturbance(VirtualMonDistBase):
             for _ in range(n_samples):
                 size = self.channels_size[channel]
                 dtype = self.channels_dtype[channel]
-                bytes = buffer[:size]
-                value = convert_bytes_to_dtype(bytes, dtype)
+                data_bytes = buffer[:size]
+                value = convert_bytes_to_dtype(data_bytes, dtype)
                 if isinstance(value, (int, float)):
                     self.channels_data[channel] = np.append(self.channels_data[channel], value)
                 buffer = buffer[size:]
@@ -1239,14 +1252,14 @@ class VirtualDrive(Thread):
     PATH_DICTIONARY_RELATIVE = "./resources/virtual_drive.xdf"
 
     def __init__(self, port: int, dictionary_path: Optional[str] = None) -> None:
-        super(VirtualDrive, self).__init__()
+        super().__init__()
         self.ip = self.IP_ADDRESS
         self.port = port
 
         self.environment = Environment()
         self.__gpios = Gpios(self.environment)
 
-        self.__register_signals: Dict[str, Signal[Any]] = {
+        self.__register_signals: dict[str, Signal[Any]] = {
             "IO_IN_VALUE": self.__gpios.value,
             "IO_IN_POLARITY": self.__gpios.polarity,
         }
@@ -1257,14 +1270,14 @@ class VirtualDrive(Thread):
         self.dictionary_path = dictionary_path or default_dictionary
         self.__stop = False
         self.device_info = None
-        self.__logger: List[Dict[str, Union[float, bytes, str, Tuple[str, int]]]] = []
-        self.__reg_address_to_id: Dict[int, Dict[int, str]] = {}
+        self.__logger: list[dict[str, Union[float, bytes, str, tuple[str, int]]]] = []
+        self.__reg_address_to_id: dict[int, dict[int, str]] = {}
         self.__dictionary = DictionaryFactory.create_dictionary(
             self.dictionary_path, Interface.VIRTUAL
         )
-        self.reg_signals: Dict[str, NDArray[np.float_]] = {}
-        self.reg_time: Dict[str, NDArray[np.float_]] = {}
-        self.reg_noise_amplitude: Dict[str, float] = {}
+        self.reg_signals: dict[str, NDArray[np.float64]] = {}
+        self.reg_time: dict[str, NDArray[np.float64]] = {}
+        self.reg_noise_amplitude: dict[str, float] = {}
         self.is_monitoring_available = EthernetServo.MONITORING_DATA in self.__dictionary.registers(
             0
         )
@@ -1278,7 +1291,9 @@ class VirtualDrive(Thread):
 
         self._monitoring: Optional[VirtualMonitoring] = None
         self._disturbance: Optional[VirtualDisturbance] = None
-        if self.__register_exists(0, VirtualMonitoring.STATUS_REGISTER):
+        if self.__register_exists(0, VirtualMonitoring.STATUS_REGISTER) and self.__register_exists(
+            0, VirtualDisturbance.STATUS_REGISTER
+        ):
             self._monitoring = VirtualMonitoring(self)
             self._disturbance = VirtualDisturbance(self)
 
@@ -1318,7 +1333,7 @@ class VirtualDrive(Thread):
             except Exception:
                 continue
             reg_add, subnode, cmd, data = MCB.read_mcb_frame(frame)
-            self.__log(add, frame, MSG_TYPE.RECEIVED)
+            self.__log(add, frame, MsgType.RECEIVED)
             register = self.get_register(subnode, reg_add)
             if cmd == self.WRITE_CMD:
                 response = self.__get_response_to_write_command(register, data)
@@ -1426,6 +1441,7 @@ class VirtualDrive(Thread):
 
     def _create_monitoring_disturbance_registers(self) -> None:
         """Create the monitoring and disturbance data registers.
+
         Only used for dictionaries V2 because V3 includes these registers.
 
         """
@@ -1437,30 +1453,30 @@ class VirtualDrive(Thread):
             ],
             EthernetServo.DIST_DATA: self.__dictionary.registers(0)[EthernetServo.DIST_DATA],
         }
-        for id, reg in custom_regs.items():
+        for register_id, reg in custom_regs.items():
             if not isinstance(reg, EthernetRegister):
                 raise ValueError
             register = EthernetRegister(
                 reg.address,
                 REG_DTYPE.BYTE_ARRAY_512,
                 reg.access,
-                identifier=id,
+                identifier=register_id,
                 subnode=reg.subnode,
             )
             self.__dictionary._add_register_list(register)
-            self.__dictionary.registers(reg.subnode)[id].storage_valid = True
+            self.__dictionary.registers(reg.subnode)[register_id].storage_valid = True
 
-            self.__reg_address_to_id[reg.subnode][reg.address] = id
+            self.__reg_address_to_id[reg.subnode][reg.address] = register_id
 
     def _init_register_signals(self) -> None:
         """Init signals, vector time and noise amplitude for each register."""
         for subnode in self.__dictionary.subnodes:
-            for reg_id in self.__dictionary.registers(subnode).keys():
+            for reg_id in self.__dictionary.registers(subnode):
                 self.reg_signals[reg_id] = np.array([])
                 self.reg_time[reg_id] = np.array([])
                 self.reg_noise_amplitude[reg_id] = 0.0
 
-    def __send(self, response: bytes, address: Tuple[str, int]) -> None:
+    def __send(self, response: bytes, address: tuple[str, int]) -> None:
         """Send a message and update log.
 
         Args:
@@ -1468,7 +1484,7 @@ class VirtualDrive(Thread):
             address: IP address and port.
         """
         self.socket.sendto(response, address)
-        self.__log(address, response, MSG_TYPE.SENT)
+        self.__log(address, response, MsgType.SENT)
 
     def _response_monitoring_data(self, data: bytes) -> bytes:
         """Creates a response for monitoring data.
@@ -1490,7 +1506,7 @@ class VirtualDrive(Thread):
         self._monitoring.available_bytes = len(data_left)
         return response
 
-    def __log(self, ip_port: Tuple[str, int], message: bytes, msg_type: MSG_TYPE) -> None:
+    def __log(self, ip_port: tuple[str, int], message: bytes, msg_type: MsgType) -> None:
         """Updates log.
 
         Args:
@@ -1508,7 +1524,7 @@ class VirtualDrive(Thread):
         )
 
     @property
-    def log(self) -> List[Dict[str, Union[float, bytes, str, Tuple[str, int]]]]:
+    def log(self) -> list[dict[str, Union[float, bytes, str, tuple[str, int]]]]:
         """Dictionary containing log information."""
         return self.__logger
 
@@ -1625,7 +1641,7 @@ class VirtualDrive(Thread):
         return self.__reg_address_to_id[subnode][address]
 
     def id_to_address(self, subnode: int, uid: str) -> int:
-        """Return address of Register with the target subnode and UID
+        """Return address of Register with the target subnode and UID.
 
         Args:
             subnode: Subnode.
@@ -1639,76 +1655,81 @@ class VirtualDrive(Thread):
             raise ValueError(f"Register {uid} is not an EthernetRegister")
         return register.address
 
-    def get_value_by_id(self, subnode: int, id: str) -> Union[int, float, str, bytes]:
+    def get_value_by_id(self, subnode: int, register_id: str) -> Union[int, float, str, bytes]:
         """Returns a register value by its ID.
 
         Args:
             subnode: Subnode.
-            id: Register ID.
+            register_id: Register ID.
 
         Returns:
             Register value.
         """
-        if id in self.__register_signals:
-            return self.__register_signals[id].get()  # type: ignore
+        if register_id in self.__register_signals:
+            return self.__register_signals[register_id].get()  # type: ignore[no-any-return]
 
-        register = self.__dictionary.registers(subnode)[id]
+        register = self.__dictionary.registers(subnode)[register_id]
         value: Union[int, float]
-        if len(self.reg_signals[id]) > 0:
+        if len(self.reg_signals[register_id]) > 0:
             actual_time = time.time()
             if self._disturbance and self._disturbance.enabled:
                 time_diff = actual_time - self._disturbance.start_time
                 actual_time = self._disturbance.start_time + (
                     time_diff % self._disturbance.buffer_time
                 )
-            sample_index = np.argmin(np.abs(self.reg_time[id] - actual_time))
-            value = self.reg_signals[id][sample_index]
-            if self.reg_noise_amplitude[id] > 0:
-                value = value + self.reg_noise_amplitude[id] * np.random.uniform()
+            sample_index = np.argmin(np.abs(self.reg_time[register_id] - actual_time))
+            value = self.reg_signals[register_id][sample_index]
+            if self.reg_noise_amplitude[register_id] > 0:
+                rng = np.random.default_rng()
+                value = value + self.reg_noise_amplitude[register_id] * rng.uniform()
 
             if register.dtype != REG_DTYPE.FLOAT:
                 value = int(value)
             return value
-        storage_value = self.__dictionary.registers(subnode)[id]._storage
+        storage_value = self.__dictionary.registers(subnode)[register_id]._storage
         if isinstance(storage_value, (int, float, str, bytes)):
             return storage_value
         else:
             return 0
 
-    def set_value_by_id(self, subnode: int, id: str, value: Union[float, int, str, bytes]) -> None:
+    def set_value_by_id(
+        self, subnode: int, register_id: str, value: Union[float, int, str, bytes]
+    ) -> None:
         """Set a register value by its ID.
 
         Args:
             subnode: Subnode.
-            id: Register ID.
+            register_id: Register ID.
             value: Value to be set.
         """
-        if id in self.__register_signals:
-            self.__register_signals[id].set(value)
+        if register_id in self.__register_signals:
+            self.__register_signals[register_id].set(value)
 
-        self.__dictionary.registers(subnode)[id].storage = value
+        self.__dictionary.registers(subnode)[register_id].storage = value
 
-    def __register_exists(self, subnode: int, id: str) -> bool:
-        """Returns True if the register exists in the dictionary
+    def __register_exists(self, subnode: int, register_id: str) -> bool:
+        """Returns True if the register exists in the dictionary.
 
         Args:
             subnode: Subnode.
-            id: Register ID.
+            register_id: Register ID.
 
         Returns:
             Register value.
         """
-        return subnode in self.__dictionary.subnodes and id in self.__dictionary.registers(subnode)
+        return subnode in self.__dictionary.subnodes and register_id in self.__dictionary.registers(
+            subnode
+        )
 
     def get_register(
-        self, subnode: int, address: Optional[int] = None, id: Optional[str] = None
+        self, subnode: int, address: Optional[int] = None, register_id: Optional[str] = None
     ) -> EthernetRegister:
         """Returns a register by its address or ID.
 
         Args:
             subnode: Subnode.
             address: Register address. Default to None.
-            id: Register ID. Default to None.
+            register_id: Register ID. Default to None.
 
         Returns:
             Register instance.
@@ -1717,13 +1738,13 @@ class VirtualDrive(Thread):
             ValueError: If both address and id are None.
         """
         if address is not None:
-            id = self.address_to_id(subnode, address)
+            register_id = self.address_to_id(subnode, address)
         else:
-            if id is None:
+            if register_id is None:
                 raise ValueError("Register address or id should be passed")
-        register = self.__dictionary.registers(subnode)[id]
+        register = self.__dictionary.registers(subnode)[register_id]
         if not isinstance(register, EthernetRegister):
-            raise ValueError(f"Register {id} is not an EthernetRegister")
+            raise ValueError(f"Register {register_id} is not an EthernetRegister")
         return register
 
     def __set_motor_enable(self) -> None:
@@ -1816,6 +1837,7 @@ class VirtualDrive(Thread):
 
     @property
     def status_word(self) -> int:
+        """Get the status word register."""
         return int(self.get_value_by_id(1, "DRV_STATE_STATUS"))
 
     @status_word.setter