+++ virtual_drive/gpios.py
@@ -7,7 +7,7 @@ if TYPE_CHECKING:
 
 
 class Gpios:
-    """General Purpose Input and Outputs"""
+    """General Purpose Input and Outputs."""
 
     def __init__(self, environment: "Environment"):
         self.value = Signal[int](0)